% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghibliapi.R
\name{get_ghibli}
\alias{get_ghibli}
\title{Retrieve data from the Ghibli API}
\usage{
get_ghibli(type = NULL, id = NULL)
}
\arguments{
\item{type}{The endpoint type, one of 'films', 'species', 'people', 'vehicles' or 'locations'.
Default to NULL if nothing is provided.}

\item{id}{The object index. Default is all unless an id is specified.}
}
\value{
Either a data.frame (if id is NULL) or a list
}
\description{
Retrieve data from the Ghibli API
}
\examples{
 get_ghibli(type="films")
 get_ghibli(type="people",id="986faac6-67e3-4fb8-a9ee-bad077c2e7fe")
 \dontrun{
 require("tidyverse")
 get_ghibli("species") \%>\%
 filter(name=="Cat") \%>\%
 select(id) \%>\%
 get_ghibli("species",id=.) \%>\%
 subset(.,str_detect(.,"people")) \%>\%
 str_replace("https://ghibliapi.herokuapp.com/people/","") \%>\%
 map(~ get_ghibli("people",.)) \%>\%
 {
  tibble(
  name = map_chr(., "name"),
  gender = map_chr(., "gender"),
  eye_color = map_chr(., "eye_color"),
  hair_color = map_chr(., "hair_color")
  )
 }
}
}
\references{
https://ghibliapi.herokuapp.com
}
