% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghibliapi-package.R
\docType{package}
\name{ghibliapi-package}
\alias{ghibliapi-package}
\alias{ghibliapi}
\title{R wrapper for the Ghibli API}
\description{
This package provides access to the \href{https://ghibliapi.herokuapp.com}{Ghibli} API from R.
}
